fingerprint_recognition_v2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
tqdm>=4.60.0
